Change algorithm for calculating sparsity
The ComputeSparsity algorithm had a bug in it was using MP_INTEGER in the
MPI_ALLREDUCE.  This resulted in numbers like 4.243991605645696E-315 for
the answer.  Also it was using logs when that wasn't necessary.
Change to just convert the rows,columns,numelemnts to reals and compute
directly.  Do a reduction on numelements if a comm is provided.