Add check for GNU10 and NAG flags
Add code from Netcdf configure.ac which will check for the
-fallow-argument-mismatch flag introduced in gfortran 10 and
the -mismatch_all flag in NAG.  These are needed because those
compilers will by default error if there is mismatch between actual and dummy
arguments and MPICH has that issue.

Also move the check for "ifort" to after "gnu" so that wrappers
like mpifort that wrap gnu are not mistaken for Intel ifort.